move getLinkDir to action
